[GENESIS] Move GetGenesisBlocksIDs from state to genesis
